# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: games
    description: Game generation and management
  - name: users
    description: User profile operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /games/generate:
    post:
      operationId: generateGame
      tags: [games]
      summary: Generate a playable browser game via AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameRequest"
      responses:
        "200":
          description: Generated game code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameGenerated"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Generation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /games/public:
    get:
      operationId: listPublicGames
      tags: [games]
      summary: List published community games
      parameters:
        - name: genre
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of published games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GameSummary"

  /games/my:
    get:
      operationId: listMyGames
      tags: [games]
      summary: List the current user's games (drafts and published)
      responses:
        "200":
          description: User's games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GameSummary"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /games:
    post:
      operationId: saveGame
      tags: [games]
      summary: Save a generated game as a draft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameInput"
      responses:
        "201":
          description: Saved draft game
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /games/{id}:
    get:
      operationId: getGame
      tags: [games]
      summary: Get a single game by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Game details with full code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateGame
      tags: [games]
      summary: Update a draft game's code or title
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GamePatch"
      responses:
        "200":
          description: Updated game
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteGame
      tags: [games]
      summary: Delete a game
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /games/{id}/publish:
    post:
      operationId: publishGame
      tags: [games]
      summary: Publish a draft game to the community gallery
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Published game
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /games/{id}/chat:
    post:
      operationId: chatEditGame
      tags: [games]
      summary: Use AI to apply a natural-language edit to a game
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameChatRequest"
      responses:
        "200":
          description: Updated game code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameChatResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Game not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: AI edit failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /games/{id}/like:
    post:
      operationId: toggleLike
      tags: [games]
      summary: Toggle like on a published game
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Like toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /games/slug/{slug}:
    get:
      operationId: getGameBySlug
      tags: [games]
      summary: Get a published game by its shareable slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Game details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user's profile and stats
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/me/liked:
    get:
      operationId: getMyLikedGames
      tags: [users]
      summary: Get game IDs liked by the current user
      responses:
        "200":
          description: Liked game IDs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: integer

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # AI generation input/output (no DB save)
    GameRequest:
      type: object
      required: [prompt, genre]
      properties:
        prompt:
          type: string
          minLength: 1
        genre:
          type: string
          enum: [Platformer, Horror, Shooter, Puzzle, Racing, RPG, Adventure, Fantasy, FP Horror, Space Shooter]
        engine:
          type: string
          enum: ["2d", "3d"]
          default: "2d"

    GameGenerated:
      type: object
      required: [gameCode, title]
      properties:
        gameCode:
          type: string
        title:
          type: string
        qualityScore:
          type: integer
          minimum: 0
          maximum: 100
          description: AI-assessed quality score (0–100) for the generated game
        gamePlan:
          type: string
          description: Game design document Claude wrote before building the game

    # Full game object (with code)
    Game:
      type: object
      required: [id, title, genre, prompt, gameCode, status, slug, likesCount, authorId, authorName, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        genre:
          type: string
        prompt:
          type: string
        gameCode:
          type: string
        currentCode:
          type: string
          nullable: true
          description: Latest live version of the code, updated after every successful AI chat edit
        codeVersion:
          type: integer
          description: Increments each time an AI chat edit is successfully applied
        coverImageUrl:
          type: string
          nullable: true
          description: Object-storage path for the AI-generated cover image
        status:
          type: string
          enum: [draft, published]
        slug:
          type: string
        likesCount:
          type: integer
        authorId:
          type: string
        authorName:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # Summary for list views (no code)
    GameSummary:
      type: object
      required: [id, title, genre, status, slug, likesCount, authorId, authorName, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        genre:
          type: string
        status:
          type: string
          enum: [draft, published]
        slug:
          type: string
        likesCount:
          type: integer
        authorId:
          type: string
        authorName:
          type: string
        createdAt:
          type: string
          format: date-time

    # Save generated game as draft
    GameInput:
      type: object
      required: [title, genre, prompt, gameCode]
      properties:
        title:
          type: string
          minLength: 1
        genre:
          type: string
        prompt:
          type: string
        gameCode:
          type: string

    # Patch existing game
    GamePatch:
      type: object
      properties:
        title:
          type: string
        gameCode:
          type: string

    GameChatRequest:
      type: object
      required: [message]
      properties:
        message:
          type: string
          minLength: 1
        currentCode:
          type: string
          minLength: 1
          description: Deprecated — server now uses DB-tracked currentCode as the authoritative source. Field accepted for backwards compatibility but ignored.

    GameChatResponse:
      type: object
      required: [updatedCode]
      properties:
        updatedCode:
          type: string
        changeSummary:
          type: string
          description: One-sentence description of exactly what Claude changed
        codeVersion:
          type: integer
          description: New version number after this edit was applied

    LikeResult:
      type: object
      required: [liked, likesCount]
      properties:
        liked:
          type: boolean
        likesCount:
          type: integer

    UserProfile:
      type: object
      required: [id, name, totalGames, publishedGames, draftGames, totalLikes]
      properties:
        id:
          type: string
        name:
          type: string
        totalGames:
          type: integer
        publishedGames:
          type: integer
        draftGames:
          type: integer
        totalLikes:
          type: integer