# lib/api-zod/src/generated/types/game.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GameStatus } from './gameStatus';

export interface Game {
  id: number;
  title: string;
  genre: string;
  prompt: string;
  gameCode: string;
  currentCode?: string | null;
  codeVersion?: number;
  coverImageUrl?: string | null;
  status: GameStatus;
  slug: string;
  likesCount: number;
  authorId: string;
  authorName: string;
  createdAt: Date;
  updatedAt: Date;
}